# perf(es/minifier): Adjust parallelism threshold