--- a/.changeset/swift-days-accept.md
+++ b/.changeset/swift-days-accept.md
@@ -1,0 +1,6 @@
+---
+swc_ecma_minifier: minor
+swc_core: minor
+---
+
+perf(es/minifier): Adjust parallelism threshold